refactor: move cms/lms module to a single api module
